+++ pybryt/__init__.py
@@ -1,6 +1,8 @@
 """Auto-assessment library for Python"""
 
+from . import utils
 from .annotations import *
+from .debug import *
 from .execution import *
 from .reference import *
 from .student import *