+++ tests/utils.py
@@ -1,8 +1,5 @@
 """"""
 
-import numpy as np
-import pandas as pd
-
 
 class AttrDict(dict):